Creación: clase pedido, prototipo
